+++ getAddress.Sdk.Standard/Api/InvoiceCCApi.cs
@@ -62,7 +62,7 @@ namespace getAddress.Sdk.Api
             };
 
             Func<string, string, AddInvoiceCCResponse> tokenExpired = (rp, b) => { return new AddInvoiceCCResponse.TokenExpired(rp, b); };
-            Func<string, string, int, AddInvoiceCCResponse> limitReached = (rp, b, r) => { return new AddInvoiceCCResponse.RateLimitedReached(rp, b, r); };
+            Func<string, string, double, AddInvoiceCCResponse> limitReached = (rp, b, r) => { return new AddInvoiceCCResponse.RateLimitedReached(rp, b, r); };
             Func<int, string, string, AddInvoiceCCResponse> failed = (sc, rp, b) => { return new AddInvoiceCCResponse.Failed(sc, rp, b); };
 
             return response.GetResponse( body,
@@ -100,7 +100,7 @@ namespace getAddress.Sdk.Api
             };
 
             Func<string, string, RemoveInvoiceCCResponse> tokenExpired = (rp, b) => { return new RemoveInvoiceCCResponse.TokenExpired(rp, b); };
-            Func<string, string, int, RemoveInvoiceCCResponse> limitReached = (rp, b, r) => { return new RemoveInvoiceCCResponse.RateLimitedReached(rp, b, r); };
+            Func<string, string, double, RemoveInvoiceCCResponse> limitReached = (rp, b, r) => { return new RemoveInvoiceCCResponse.RateLimitedReached(rp, b, r); };
             Func<int, string, string, RemoveInvoiceCCResponse> failed = (sc, rp, b) => { return new RemoveInvoiceCCResponse.Failed(sc, rp, b); };
 
             return response.GetResponse(body,
@@ -136,7 +136,7 @@ namespace getAddress.Sdk.Api
             };
 
             Func<string, string, ListInvoiceCCResponse> tokenExpired = (rp, b) => { return new ListInvoiceCCResponse.TokenExpired(rp, b); };
-            Func<string, string, int, ListInvoiceCCResponse> limitReached = (rp, b, r) => { return new ListInvoiceCCResponse.RateLimitedReached(rp, b, r); };
+            Func<string, string, double, ListInvoiceCCResponse> limitReached = (rp, b, r) => { return new ListInvoiceCCResponse.RateLimitedReached(rp, b, r); };
             Func<int, string, string, ListInvoiceCCResponse> failed = (sc, rp, b) => { return new ListInvoiceCCResponse.Failed(sc, rp, b); };
 
             return response.GetResponse(body,
@@ -167,7 +167,7 @@ namespace getAddress.Sdk.Api
             };
 
             Func<string, string, GetInvoiceCCResponse> tokenExpired = (rp, b) => { return new GetInvoiceCCResponse.TokenExpired(rp, b); };
-            Func<string, string, int, GetInvoiceCCResponse> limitReached = (rp, b, r) => { return new GetInvoiceCCResponse.RateLimitedReached(rp, b, r); };
+            Func<string, string, double, GetInvoiceCCResponse> limitReached = (rp, b, r) => { return new GetInvoiceCCResponse.RateLimitedReached(rp, b, r); };
             Func<int, string, string, GetInvoiceCCResponse> failed = (sc, rp, b) => { return new GetInvoiceCCResponse.Failed(sc, rp, b); };
 
             return response.GetResponse( body,
