+++ getAddress.Sdk.Standard/Api/PermissionApi.cs
@@ -50,7 +50,7 @@ namespace getAddress.Sdk.Api
             };
 
             Func<string, string, PermissionResponse> tokenExpired = (rp, b) => { return new PermissionResponse.TokenExpired(rp, b); };
-            Func<string, string, int, PermissionResponse> limitReached = (rp, b, r) => { return new PermissionResponse.RateLimitedReached(rp, b, r); };
+            Func<string, string, double, PermissionResponse> limitReached = (rp, b, r) => { return new PermissionResponse.RateLimitedReached(rp, b, r); };
             Func<int, string, string, PermissionResponse> failed = (sc, rp, b) => { return new PermissionResponse.Failed(sc, rp, b); };
 
             return response.GetResponse( body,
@@ -85,7 +85,7 @@ namespace getAddress.Sdk.Api
             };
 
             Func<string, string, ListPermissionResponse> tokenExpired = (rp, b) => { return new ListPermissionResponse.TokenExpired(rp, b); };
-            Func<string, string, int, ListPermissionResponse> limitReached = (rp, b, r) => { return new ListPermissionResponse.RateLimitedReached(rp, b, r); };
+            Func<string, string, double, ListPermissionResponse> limitReached = (rp, b, r) => { return new ListPermissionResponse.RateLimitedReached(rp, b, r); };
             Func<int, string, string, ListPermissionResponse> failed = (sc, rp, b) => { return new ListPermissionResponse.Failed(sc, rp, b); };
 
             return response.GetResponse( body,
@@ -120,7 +120,7 @@ namespace getAddress.Sdk.Api
             };
 
             Func<string, string, UpdatePermissionResponse> tokenExpired = (rp, b) => { return new UpdatePermissionResponse.TokenExpired(rp, b); };
-            Func<string, string, int, UpdatePermissionResponse> limitReached = (rp, b, r) => { return new UpdatePermissionResponse.RateLimitedReached(rp, b, r); };
+            Func<string, string, double, UpdatePermissionResponse> limitReached = (rp, b, r) => { return new UpdatePermissionResponse.RateLimitedReached(rp, b, r); };
             Func<int, string, string, UpdatePermissionResponse> failed = (sc, rp, b) => { return new UpdatePermissionResponse.Failed(sc, rp, b); };
 
             return response.GetResponse(body,
@@ -155,7 +155,7 @@ namespace getAddress.Sdk.Api
             };
 
             Func<string, string, AddPermissionResponse> tokenExpired = (rp, b) => { return new AddPermissionResponse.TokenExpired(rp, b); };
-            Func<string, string, int, AddPermissionResponse> limitReached = (rp, b, r) => { return new AddPermissionResponse.RateLimitedReached(rp, b, r); };
+            Func<string, string, double, AddPermissionResponse> limitReached = (rp, b, r) => { return new AddPermissionResponse.RateLimitedReached(rp, b, r); };
             Func<int, string, string, AddPermissionResponse> failed = (sc, rp, b) => { return new AddPermissionResponse.Failed(sc, rp, b); };
 
             return response.GetResponse( body,
@@ -193,7 +193,7 @@ namespace getAddress.Sdk.Api
             };
 
             Func<string, string, RemovePermissionResponse> tokenExpired = (rp, b) => { return new RemovePermissionResponse.TokenExpired(rp, b); };
-            Func<string, string, int, RemovePermissionResponse> limitReached = (rp, b, r) => { return new RemovePermissionResponse.RateLimitedReached(rp, b, r); };
+            Func<string, string, double, RemovePermissionResponse> limitReached = (rp, b, r) => { return new RemovePermissionResponse.RateLimitedReached(rp, b, r); };
             Func<int, string, string, RemovePermissionResponse> failed = (sc, rp, b) => { return new RemovePermissionResponse.Failed(sc, rp, b); };
 
             return response.GetResponse( body,
