+++ getAddress.Sdk.Standard/Api/Responses/ListDomainWhitelistResponse.cs
@@ -52,8 +52,8 @@ namespace getAddress.Sdk.Api.Responses
 
         public class RateLimitedReached : Failed
         {
-            public int RetryAfterSeconds { get; }
-            public RateLimitedReached(string reasonPhrase, string raw, int retryAfterSeconds) : base(429, reasonPhrase, raw)
+            public double RetryAfterSeconds { get; }
+            public RateLimitedReached(string reasonPhrase, string raw, double retryAfterSeconds) : base(429, reasonPhrase, raw)
             {
                 RetryAfterSeconds = retryAfterSeconds;
                 RateLimitReachedResult = this;
